Added world properties to Format class and deferred loading
When a Format class instance is created with a directory it will initially only load the level.dat file and parse data from it.
This will expose the level name and game version
At the first point data is requested from the Format class (getting chunks, setting chunks, deleting chunks, saving, ...) the world data will be loaded.
This allows a UI to have easy access to data about the world without loading the whole world