Restructured java interfaces and redesigned is_valid
Java interfaces are now primarily named after the data version they are first used in.
This is because one unique format may stretch across the major version numbers or multiple interfaces may be needed per major version.
Also slightly redesigned is_valid to be easier to work with. 'java' is now tested in the base class so it doesn't need checking in each