Fixed SubBox so the boundaries are the actual boundary locations
It used to behave like the Java fill command where a range from 0 to 15 would fill the whole chunk which is unintuative for python. It has now been rewritten so that 0 to 16 is the whole chunk like how the range function works in python.
This is more intuative because a box from 0 to 6 and a box from 5 to 10 should intersect