Added full support for mcc files
Previous to 1.15 if a chunk was greater than 1MiB it just wasn't saved. 1.15 added the mcc file which allows chunks of greater size to be saved without this limitation.
This implementation should only read and write them if the world dataversion is greater than 2203 which is about when it was first released.