Made selection compulsory in get_chunk_boxes
I wrote out most of the logic to iterate over chunks rather than boxes if the boxes got too large to optimise things a bit.
I have now scrapped this as there were too many issues with it.
The simple answer is this method isn't designed to be used in cases where the selection might be the whole world.