Reimplemented iterating over the existing chunks
Previously this method got a box per sub-chunk in the box. This meant that if the box was very tall it would still mean iterating over a large range.
I have now switched it so that each box and chunk combination will only yield one box rather than one per sub-chunk.
The previous logic of iterating over existing chunks and checking if they are in the selection is viable for large selections.