Split get_chunk_boxes into two methods
There are some methods that need to know the progress of get_chunk_boxes but in its previous state the only way was to get its length. This meant accessing all the chunks twice. The new way splits it into a method that doesn't access the chunks and one that does access the chunks