Added more locks to Java region
Previously locks were added to the Java region code to fix some chunk loading issues with threads.
This extends that to encase all code that modifies the state of _chunks in locks.
There was a case where the region file was being loaded and then unloaded in another thread before the first had a chance to access the data leading to a None type is not iterable error.