Added purge methods to unload all data and reload from disk.
This enables switching between programs that manipulate the higher level data and programs that edit the low level data as it is on disk.
The class will need to be saved if desired and then this method called to clear all of the cache data.
When switching back instead of repopulating from the cached data it will be populated from the data on disk like it was never populated.