Removed run_operation method
This is never used in the map editor codebase so it may confuse users.
It is also too high level for users of the core library who should have the option to create an undo point.
The preferred way of doing this behaviour is to directly call the function giving the function the level object.