Cleaned up the database logic
Moved some of the methods from the ChunkManager class into the database class.
Extended the get_entry behaviour to the has_entry and all_entries methods.
The old has and all methods just looked at the data in the class itself. Wrapping classes then had to handle getting the data from the level if it was not in the class.
This modifies that so that the base method will call a method to request it from the level if required.
This should simplify the logic a little.
This still needs implementing for the player logic.