Cleaned up the chunk get methods
The try except logic is required but was a bit bulky to duplicate a number of times.
This moves it to a helper method where it can be reused.
This logic has been added to the raw get method.
get_raw_chunk_data is the public method so also includes this method. The load_chunk method now uses the private one so the try except logic is not run twice.