Cleaned up the database logic (#132)
* Fixed some minor issues

* Cleaned up the database logic

Moved some of the methods from the ChunkManager class into the database class.
Extended the get_entry behaviour to the has_entry and all_entries methods.
The old has and all methods just looked at the data in the class itself. Wrapping classes then had to handle getting the data from the level if it was not in the class.
This modifies that so that the base method will call a method to request it from the level if required.
This should simplify the logic a little.
This still needs implementing for the player logic.

* Added some documentation about the database

* Added the raw setup to the player manager

This makes some methods abstract so they will need implementing in all subclasses.
It also adds a new method which will also need implementing

* Removed incorrect abstractmethod decorator

* Fixed typing

* Renamed the level all players method

* Added custom errors and cleaned up typing

* Switched all_chunk_coords typing to use iterable

* Switched _uuid to uuid in Player init and added type checking

* Cleaned up the chunk get methods

The try except logic is required but was a bit bulky to duplicate a number of times.
This moves it to a helper method where it can be reused.
This logic has been added to the raw get method.
get_raw_chunk_data is the public method so also includes this method. The load_chunk method now uses the private one so the try except logic is not run twice.

* Reworked the format wrapper player methods to match the chunk methods

Added methods to access the low level player data.
Renamed get method to load to better describe what it does.
Modified the leveldb and java methods to use this new format.
Added no-op methods for the other classes since they don't have player data.

* Removed unused import

* Force leveldb inputs to be bytes

* Added contains and iter methods to leveldb

Also added more tests

* Added has_player methods

* Fixed player tests

* Removed use of depreciated input