Finished up implementation and merged into one class
I realised the multiple class situation was just making the situation more complex than it needed to be.
The biome class now holds all the data and exposes methods to access and modify the data.
The class now doesn't store the dimension state. It is stored in both simultaneously.
The null state is handled by the data being None or empty.
The biome array can also technically be any shape between 1 and 16.