Improved the interfaces setup
The old system of suppressing behaviour was good on the small scale and was good because it kept all the logic together but it was starting to get difficult to manage especially with the new chunk structure.
This rewrite will remove the feature system and just have new behaviour overwrite the old methods.
If behaviour needs to be removed the encode/decode method will need duplicating and modifying.