Populate lower biome arrays
The game will crash if any of the biome arrays below the highest is not defined or there are no biome arrays.
This will populate the data if it does not exist so that the game does not crash.
This should fix https://github.com/Amulet-Team/Amulet-Map-Editor/issues/475 and https://github.com/Amulet-Team/Amulet-Map-Editor/issues/491 and part of https://github.com/Amulet-Team/Amulet-Map-Editor/issues/490