Fixed player rotation being NaN
This forces player position and rotation to be in the range +/-100M
If the player rotation is NaN this causes other issues.
Also handles the case that position or rotation is not defined.
This issue was only noticed in Java but making the changes to Bedrock just in case.
Fixes https://github.com/Amulet-Team/Amulet-Map-Editor/issues/448