Added support for multiple data layers in Java edition
In 1.17 Java edition split the entity data from the rest of the region data.
This keeps the behaviour the same but adds get_chunk_data_layers and put_chunk_data_layers which applies to all layers of the data.
This will enable loading entity data but won't break old code.