Rewritten the java decoder logic
The chunk decoding and encoding logic has gone through a few incarnations.
To begin with all the logic was in one function with feature toggles to change the behaviour.
The second version was individual methods which were called from one method. This had the disadvantage that to disable one feature the calling method needed redefining.
The new method is like the second method but the methods can be individual enabled and disabled.
The rewrite also makes it so that each feature is independent which makes it easier to disable individual features.