Improved logging support
Rather than having an object that needs importing leading to circular imports this switches to the `logging.getLogger` method.
It also cleans up the logging setup code.

Due to the nature of amulet_core being a library that can be used through an application and directly in scripting it is advantageous to have a default handler and a way for applications to specify their own handlers.
The default handler gives feedback to scripters so that they don't need to set up logging support and a function can be used to disable the default handlers.

Currently there is a file handler but this will be removed in the future when implemented in the main application.
`factory_reset` should be called after importing but before setting up custom handlers in an application.