Improved Java edition locking
Before it was possible to open a level in game and in Amulet at the same time. The first process to access it would silently loose the lock and stop saving.
Added system locks to the session.lock file so that it will error upon opening if open elsewhere.
The game will also display a warning that it is open elsewhere.