Refactored Bedrock dimension class
This previously stored data and wrote it in one go as part of the save method.
It now directly applies the changes to the database to reduce memory usage.
Added a thread lock to stop multiple operations changing the state at the same time.
Simplified some of the logic a bit