Added support for reading and writing actor keys
This should be moved into the interface however currently the interfaces are designed to have no knowledge beyond the data of the chunk.
It might make sense to rewrite the interfaces to know about the world.
At the very least the interfaces should be split so that each format can be different.