Improved the loader logic
Removed pyinstaller workaround that is no longer needed.
Changed loading logic to stop searching for extensions if the package __init__.py file contains an extension. If the package is an extension it should not contain sub-extensions. This is required to be able to move interface extensions into the format module.