Moved the Bedrock interfaces into the leveldb format
There was no good reason to have these grouped together.
It makes far more sense for each format wrapper to control its interfaces and allows the interfaces to become more specialised for each format wrapper.
It also means that finding interfaces will be quicker because is_valid does not need to be called on each interface for each chunk to find an iterface that can handle the chunk.
Moved the chunk version attribute to a class variable.