Fixed some issues with leveldb worlds
The previous code was running the repair logic each time the world was loaded.
This was creating a lost directory with an increasing amount of data.
The new implementation only runs the repair logic if the world did not open correctly.
If opening raised an OSError (usually marketplace worlds) a new distinct error is raised.
Repair could partially fix the marketplace leveldb but not fully causing chunks to get deleted.
This reverts it so that marketplace worlds always throw an error when loading.