Replaced global cache db with a per level db
The global cache was problematic if multiple processes were spawned because they would all try and open the same cache directory.
This change removed the global cache location and adds a cache per level.
The cache is now stored in the system temp directory as found using the tempfile module.
A custom class has been added to handle automatic deletion of the cache directory when the instance is lost.
Worst case if the program crashes they will be deleted after a week if the lock file is still not being used.