Added classmethod constructors to FormatWrapper
This was originally designed to only open existing data and loading was done in the constructor.
This had a number of design problems with subclassing and creating instances without existing data.
The constructor now only defines default attributes and does not load any data.
create_and_open is mostly the same as it was before but is now a classmethod meaning it doesn't require an instance.
Added load classmethod to replace the old constructor.