Split the create and load abstract methods
Not all data sources are both creatable and loadable so it doesn't make sense to enforce both on all.
Having the create and load abstract methods in different classes allows the concrete classes to inherit as required making it easier to tell which format wrappers are creatable and which are loadable.