Added a global cache
All levels will share one cache class.
It would be difficult to synchronise multiple cache classes not to fill a user's RAM if the cache size is set large.
Having one global cache allows multiple levels to share one large cache to avoid this issue.

Added a del method to cleanup when the class is destroyed.