Split the history manager into two classes
All methods that interact with the data from one layer are now in a new class.
This means that the arbitrary layer key no longer exists. The uuid for the layer now takes this job.
Moved the data into a private class so it can be shared by both classes.
The resource id is no longer required to be a bytes object. It must match the ResourceID protocol and can define its own attributes.