Added opened and closed signals to the private data
This will allow friends of the level to get notified when the level opens and closes so they can initialise and discard data.
The alternative was to expose methods on the friends that the level could call but these would be public to third party code so this seems like a better solution.