Refactored the Block class
Added a Immutable and Hashable properties class
Split the Block class into a Block class and BlockStack class. The old block class was theoretically infinitely recursive. In the new implementation the Block class represents a single block state and the BlockStack class represents a stack of one or more blocks.
Removed some methods