Refactored private data handling
I was trying to create a class storing private data for a class which could be shared with friends of the class but it was becoming difficult to manage.
I have now rolled all the data back into the original classes as private attributes which friends can access directly.
All data defined at creation is normal attributes and all other data defined when the level is opened should be defined in one attribute to make error handling easier