Renamed translate to game
This module holds information about each game version as well as being able to translate between them.
As such this feels like a more fitting name.
Removed TranslationManager class. This class previously stored modifications to the data for the level it was associated with.
I now think we should have a single instance for each version that is immutable and the level implementation can build upon it