Added compiler metadata and validator
The libraries must be compiled with the same version of the compiler and pybind11.
If they are compiled with different versions bad things happen.
Adding this data at compile time allows dependent libraries to validate compatibility.

Added code to validate that the nbt and leveldb libraries are compatible.
This produces a sane error message at import.