Removed compiled builds
The libraries all need to be compiled with the same compiler and same version of pybind11.
Pypi and pip do not have a mechanism to manage this so it would be a nightmare.
I would like to support compiled builds but I can't figure out a way to ensure they are all built by the same compiler.