Removed RegionFile closer class
This class made the whole system more complex.
It was designed with the intention to not destroy RegionFile instances which was a bad decision.
The file object will now be kept open until the region instance is destroyed or close called.
A limited number of region instances should be kept alive per level to minimise resources used.