Redesign selection classes (#422)
* Update stubs

* Rename SelectionGroup to SelectionBoxes

SelectionGroup is an ambiguous name for a sequence of SelectionBox instances.
The new name is a better name for what it actually is.
The old name has been kept as an alias but may be removed in the future.

* Replace SelectionBoxes size with count

size is ambiguous. It could relate to the shape of the box.

* Rename SelectionBoxes to SelectionBoxGroup

* Update pybind11 class type

* Add SelectionShape

This is a base class for all selection shape classes.

* Rename missed SelectionBoxes names

* Redesign SelectionBoxGroup constructors

* Add iterators to SelectionBoxGroup

* Add cast operator to SelectionBoxGroup

* Add SelectionShapeGroup

This can contain zero or more arbitrary shapes for higher level manipulation

* Add SelectionSphere class

* Improve python wrapper for SelectionBoxGroup

* Add virtual shape methods

* Improve sphere voxelisation

* Remove SelectionShape from SelectionBox

* Add constructor

* Split selection module

* Update stubs

* Remove group header

* Add utils dependency

* Refactor selection shape classes

All shapes now have a matrix to allow full transforms.
SelectionSphere is now SelectionEllipsoid

* Simplify selection group constructors

* Change volume data type

* Refactor operators

* Improve comments

* Add missing namespace

* Add cuboid shape

* Fix docstring

* Refactor selection tests

* Add transform methods to box and box group

* Add transform bindings

* Add initialiser list constructor to SelectionBoxGroup

* Remove SelectionBoxGroup::size

* Rename selection box getter

* Fix 90 degree check

* Fix cuboid transformation

* Add box tests

* Add box group tests

* Add cuboid tests

* Make selection shape matrix immutable

* Implement repr for selection shapes

* Fix docstring

* Return copy of matrix

The matrix is const but pybind casts away const making it mutable on the python side.
Creating a copy is the simplest solution to this issue.

* Add ellipsoid tests

* Add shape group tests

* Bump utils library

* Improve build

* Remove ubuntu tests

* Add header