Improve selection classes (#423)
* Add casting constructors

* Add repr methods to stubs

* Convert matrix to property in stubs

* Add shape group voxelise bindings

* Add SelectionShapeGroup::almost_equal

* Add argument name

* Fix repr

* Add equality methods

These are not super useful because they only return true if they are exactly equal but they are here for completeness.

* Add selection serialisation

* Make selection matrix mutable

I was initially trying to make the selection classes immutable but have decided to make them mutable and return copies where the data needs to be immutable.

* Redesign shape group

The shape group is now mutable and stores shared pointers to allow the vector to be resized without breaking the link to python.
Redesign constructors.
Add copy constructor.
Add deep copy constructor.
Add some mutating methods to python bindings.

* Bump mypy min python version

* Bump extensions dependency

* Add MutableSequence to shape group

* Update shape group tests

* Improve stub generator

* Improve stubs

* Add default shape constructors

* Fix selection group constructor

* Remove shapes property

* Add selection shape group tests

* Remove pybind11 extensions from dev requirements

* Reformat

* Bump utils requirement